Improve AsyncQueue behavior, add new extensions to gfx::Context, minor gfx and util cleanups

+++ libraries/lib_awn_win32/include/awn/gfx/gfx_texturesamplermanager.vk.hpp
@@ -53,7 +53,7 @@ namespace awn::gfx {
                 constexpr ~TextureNode() {/*...*/}
             };
         public:
-            using TextureHandleTable = vp::util::FixedAtomicIndexAllocator<u32, Context::cTargetMaxTextureDescriptorCount>;
+            using TextureHandleTable = vp::util::FixedAtomicIndexAllocator<u16, Context::cTargetMaxTextureDescriptorCount>;
             using SamplerHandleTable = vp::util::FixedAtomicIndexAllocator<u16, Context::cTargetMaxSamplerDescriptorCount>;
             using SamplerMap         = vp::util::IntrusiveRedBlackTreeTraits<SamplerNode, &SamplerNode::rb_node>::Tree;
         private:
